REFACTORING: Moved files from last PR (NVI-Percentage controlled numbers) (#8377)
* Moved PercentageWithIcon from _atoms to _molecules since it consists of several elements

* Moved table rows to new location for nvi tables

* Moved NavigationAccordion

* Moved admin table helpers

* Moved tasks-navigation-accordion
