+++ src/components/_molecules/PercentageWithIcon.tsx
